Add implementation of `AbstractRepositoryBackend` using AWS S3
The `AwsS3RepositoryBackend` class implements `AbstractRepositoryBackend`
of `aiida-core`. This will be able to be used by a `StorageBackend`
implementation to use AWS S3 as the provider for the storage of binary
blobs.

The `boto3` library is used to create the REST API calls to the AWS S3
service. For unit testing, the `moto` library is used to mock the S3
client such that the actual service doesn't have to be hit.