Add implementation of `StorageBackend` using `AwsS3RepositoryBackend`
The `PsqlAwsS3Storage` implements the `StorageBackend` interface which
allows it to be used as a storage backend for an AiiDA profile. It
subclasses the `PsqlDosBackend` implementation that ships with
`aiida-core` which provides the bulk of the functionality. It merely
overrides the `get_repository` method, to return an instance of the
`AwsS3repositoryBackend`.