CI: Add Github Action workflow to run against real AWS S3 instance
The `aws_s3` session fixture which is automatically used, ensure that
all AWS S3 client connections are mocked using `moto` unless the
`AIIDA_S3_MOCK_AWS_S3` environment variable is set to `True`. This will
ensure that by default the AWS S3 client is mocked, but it is possible
to run them with a real `boto3` client to actually test against an AWS
S3 instance.

The credentials are likely mocked unless `AIIDA_S3_MOCK_AWS_S3` is set
to `True`, in which case the credentials are read from the variables:

 * `AWS_ACCESS_KEY_ID`
 * `AWS_SECRET_ACCESS_KEY`
 * `AWS_REGION_NAME`

If a specific bucket needs to be used, set the environment variable:

 * `AWS_BUCKET_NAME`

or a randomly generated bucket name will be created. Be careful that
the bucket will be emptied and deleted at the end of the session so only
use a bucket that is meant for testing.

The new GHA workflow `aws-s3.yml` uses the infrastructure to run the
test suite unmocked. This can be run occassionally to ensure that the
plugin actually works in production environments.