Fixtures: Modularize the fixtures creating AiiDA instance and profile
The test fixtures that are improved by modularizing them. The
`aiida_profile_factory` fixture takes a dictionary that will be merged
into the default configuration. The `config_psql_dos` and
`config_psql_aws_s3` fixtures return a profile configuration dictionary
that can be used for the `core.psql_dos` and `s3.psql_aws_s3` storage
backend plugins. They each create the resources they require, such as a
PostgreSQL instance or an AWS S3 connection, or mock it when necessary.

The `aiida_profile` fixture is an easy shortcut which will automatically
call `aiida_profile_factory` with the config of `config_psql_dos`
providing a fully isolated and temporary AiiDA instance with a profile
configured with the `core.psql_dos` backend.