Tests: Use the new modular pytest fixtures of `aiida-core`
So far, the plugin package defined its own pytest fixtures to setup a
temporary AiiDA test instance with a profile configured for the custom
`PsqlAwsS3Storage` storage backend plugin. This was necessary since the
`aiida_profile` fixture provided by `aiida-core` hardcoded the test
profile to use the `PsqlDosStorage`.

Now though, `aiida-core` provides fixture that are more configurable and
allow us to define the storage backend to use for the test profile.
Since they haven't been released yet, for the time being, we need to
install `aiida-core` from the `main` branch.

The custom fixtures are removed and replaced with two simple ones:

 * `config_psql_aws_s3`
 * `psql_aws_s3_profile`

The `config_psql_aws_s3` fixture provides a complete profile
configuration for use with the `PsqlAwsS3Storage` storage. This is used
by the `psql_aws_s3_profile` fixture which constructs and loads an
actual profile with this configuration. This fixture can then be used by
tests for the `PsqlAwsS3Storage` storage plugin.