`PsqlAwsS3Storage`: Implement the `migrator` class
The `PsqlAwsS3Storage` subclasses the `PsqlDosStorage` storage backend
provided by `aiida-core` and simply replaces the disk-objectstore for a
repository that uses an AWS S3 container.

The `StorageBackend` base class interface of `aiida-core` recently
changed and added the `initialise` method. This method should initialise
a new storage instance such that it is ready for use. The optional
argument `reset` when set to `True` should cause the storage instance to
first be destroyed before being reinitialised.

This functionality is implemented by the `PsqlDosMigrator` class for the
`PsqlDosBackend`. This class is subclassed such that the repository
components can be updated to use the `AwsS3RepositoryBackend` interface
instead of the `DiskObjectStoreRepositoryBackend` of `aiida-core`.

The `initalise(reset=True)` call will be made by `aiida_profile_factory`
when the storage of the test profile needs to be initialised. The reset
is necessary since part of the storage, such as the Postgres cluster
and database, could have been used by another storage backend in
preceding tests.