Add support for Azure Blob Storage
The `PsqlAzureBlobStorage` implementation of the storage backend is
added with the `s3.psql_azure_blob` entry point. It allows to use a
container on Azure Blob Storage as the file repository. The connection
is setup using a connection string which can be obtained from the Azure
portal for the Blob Storage service.