Tests: Skip the Azure Blob storage tests if they are mocked
Currently it is not yet possible to successfully mock the Azure Blob
storage client and so the tests will fail if no proper credentials are
provided.

Therefore, the relevant test modules are marked with the marker
`skip_if_azure_mocked` which will skip the test if the fixture
`should_mock_azure_blob` return `True`. This is always the case, unless
credentials are specified through the corresponding environment
variables.

Once a successful mocking method is developed this marker can be
removed.