Add a command line interface
This will allow to implement an easy way for users to create a new
profile with one of the storage backends provided by this plugin. The
current state of `verdi` does not allow to easily implement this since
the options of `verdi setup` are hardcoded for the `psql_dos` storage
backend. If this plugin becomes successful, we can port the
functionality of dynamically generated CLI commands to setup various
storage backends to `aiida-core`.