"file exist" operation now is performed with boto3's head_object
before it was using list_object, that by default returns 1000 objects at maximum.
The requested (existing) key may not be in the returned list.
In any case the maximum number of objects must be known in advance,
and listing all objects is a waste of resources.

solves https://github.com/sphuber/aiida-s3/issues/11