Add the `S3RepositoryBackend` base class (#13)
The S3 protocol has become so widely adopted that its interface has
become a standard. Besides the implementation of AWS, there are now open
source implementations that implement the same interface, for example
minIO (https://min.io/).

The existing implementation of the `AwsS3RepositoryBackend` class could
be used for most of these other S3 implementations as long as the
endpoint URL of the server can be specified.

Therefore the `AwsS3RepositoryBackend` implementation is moved to the
`S3RepositoryBackend` which now serves as the base class. It adds the
`endpoint_url` configuration parameter and removes the `region_name`,
which is an AWS S3 specific option.

The other repository interface implementations now simply subclass the
`S3RepositoryBackend` class and modify what is needed. The same goes for
the storage classes: the `PsqlS3Storage` class now serves as a base
implementation that is subclassed by adapters for AWS S3 and Azure Blob
Storage.

The new `S3RepositoryBackend` is tested in the CI workflow against a
minIO service. Note that the `minio/minio:edge-cicd` tag had to be used
instead of the base image since for the latter the connection to the
service always failed, for an unknown reason.