`S3RepositoryBackend`: Do not rely on `list_objects` for `open` (#14)
The `S3RepositoryBackend.open` method calls the super, which checks that
the requested file exists. It does so by calling `has_object` which in
turns calls `list_objects`. Not only is the current implementation of
`list_objects` incorrect, this requires the server to list all objects
that are stored, which is typically an expensive operation for an object
store.

Instead of checking whether the key is part of all existing keys, we
simply perform the call and catch the error returned by the server in
case the file doesn't exist. This is a lot more efficient.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>