`S3RepositoryBackend`: Fix bug in `list_objects` being incomplete (#18)
The implementation of the `list_objects` method is incorrect since it
only performs a single call and the `boto3` client by default only
returns the first 1000 keys and pagination should be used to retrieve
all objects for buckets that contain more objects than this page size.

The `list_objects` method is replaced with `list_objects_v2` since the
latter is essentially deprecated and just kept for backwards
compatibility. The implementation now also uses a paginator to yield the
keys from all objects.