Remove custom command line interface
The package shipped with its own CLI `aiida-s3`. The only function was
to provide a command that allowed creating new AiiDA profiles using any
of the storage backends. As of `aiida-core==2.5`, the command
`verdi profile setup` was added to `verdi` that dynamically creates a
subcommand for each storage plugin that is registered.

The requirement is updated to `aiida-core~=2.5`. The custom CLI is
removed entirely. The storage plugins remove the `_get_cli_options`
methods and adds the new pydantic model definition that replaces it.